Replace a list of species

Replace a list of species

ReplaceListSpecies(userid, accesstoken, listid, speciesobj)

Arguments

userid
A valid gmail address of the user

accesstoken
A valid gmail address of the user

listid
An integer id of the list to be modified

speciesObj
A species object to replace with

Value

A list with the old species and new species list

See also

https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md

Examples

  userid = "abusalehmdtayeen@gmail.com"
  accesstoken = "ya29..zQLmLjbyujJjwV6RVSM2sy-mkeaKu-9"
  listid = 12
  speciesObj = list( list(family="",scientific_name="Aix sponsa",scientific_name_authorship="", vernacular_name="Wood Duck",phylum="",nomenclature_code="ICZN",order="Anseriformes",class="") )
  ReplaceListSpecies(userid, accesstoken, listid, speciesObj)

#> $user_id
#> [1] "abusalehmdtayeen@gmail.com"
#> 
#> $date_modified
#> [1] "2017-07-05T16:29:37.398640"
#> 
#> $status_code
#> [1] 200
#> 
#> $old_species
#> $old_species[[1]]
#> $old_species[[1]]$family
#> [1] ""
#> 
#> $old_species[[1]]$scientific_name
#> [1] "Aix sponsa"
#> 
#> $old_species[[1]]$scientific_name_authorship
#> [1] ""
#> 
#> $old_species[[1]]$order
#> [1] "Anseriformes"
#> 
#> $old_species[[1]]$vernacular_name
#> [1] "Wood Duck"
#> 
#> $old_species[[1]]$phylum
#> [1] ""
#> 
#> $old_species[[1]]$nomenclature_code
#> [1] "ICZN"
#> 
#> $old_species[[1]]$class
#> [1] ""
#> 
#> 
#> 
#> $new_species
#> $new_species[[1]]
#> $new_species[[1]]$family
#> [1] ""
#> 
#> $new_species[[1]]$scientific_name
#> [1] "Aix sponsa"
#> 
#> $new_species[[1]]$scientific_name_authorship
#> [1] ""
#> 
#> $new_species[[1]]$class
#> [1] ""
#> 
#> $new_species[[1]]$vernacular_name
#> [1] "Wood Duck"
#> 
#> $new_species[[1]]$phylum
#> [1] ""
#> 
#> $new_species[[1]]$nomenclature_code
#> [1] "ICZN"
#> 
#> $new_species[[1]]$order
#> [1] "Anseriformes"
#> 
#> 
#> 
#> $list_id
#> [1] 12
#> 
#> $list_title
#> [1] "Bird Species List"
#> 
#> $message
#> [1] "Success"
#>